api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-cosmos==4.7.0
httpx==0.27.0
pydantic==2.7.4
azure-ai-projects==1.0.0b4
azure-ai-agents==1.0.0b2
